move file utils to src
