+++ frontend/src/pages/Connexion.jsx
@@ -8,7 +8,7 @@ import axios from "axios";
 import "../style_pages/Connexion.css";
 import cardApplicant from "../assets/images/card_applicant.png";
 import cardCompany from "../assets/images/card_company.png";
-import ValidationConnexion from "./utils/ValidationConnexion";
+import ValidationConnexion from "../utils/ValidationConnexion";
 
 function Connexion() {
   const [values, setValues] = useState({
