+++ frontend/src/utils/ValidationConnexion.jsx
